Improve source folder synchronization
The source folder updater used to keep source folders that were not
defined in Gradle. This could lead to a situation where the user defined
a folder that is in conflict with the Gradle model and can now no longer
synchronize. The user then needs to know how to remove the source folder
manually before the synchronize will work again.

The updater now removes folders that were not defined in Gradle. This
makes sure the Eclipse project matches the Gradle configuration.